Common: Fixed ByteCountChunk and BitCountChunk str() data field

+++ src/inet/common/packet/chunk/ByteCountChunk.cc
@@ -124,7 +124,7 @@ void ByteCountChunk::doRemoveAtBack(b length)
 std::string ByteCountChunk::str() const
 {
     std::ostringstream os;
-    os << "ByteCountChunk, length = " << B(getChunkLength()) << ", data = " << data;
+    os << "ByteCountChunk, length = " << B(getChunkLength()) << ", data = " << (int)data;
     return os.str();
 }
